[Refactor] ItemsList style 분리
